MOSTLY finished all functions
I mostly finished all functions.
We still need to figure out how to test for sdoemthing to be empty (ex.
test if the IOQ is empty, or if the LTQ is empty). These should be
functions we create, or we could just check to see if the end array
elements for these things are populated or not.
I also added the other miscellaneous functions to the bottom
(incomplete function).

if you could go through my code and look at the comments for the
incomplete code that would be great. Maybe you can fill in my missing
pieces. it probably just checking array positions to see if an elements
is actually occupied or not.

I also noticed in the variables.h you set our true false values to data
type integer. I changed them to bools as you will see.